Use V3.1 content digest in V4.1 signature
The content digest algorithm is determined by the number of bits in
the signing key to ensure the digest is not weaker than the key.
When the V3.1 scheme is used to sign an APK, it is possible for the
key from the V3.0 signature to meet the requirements for the
CHUNKED_SHA256 digest algorithm while the key from the V3.1
signature meets the requirements for the CHUNKED_SHA512 digest
algorithm. This commit ensure that the V4.1 signature uses the
content digest from the V3.1 signing block since this is the
digest that will be used to verify the V4.1 signature on all
platform versions that support the V3.1 signature.

Bug: 282777109
Test: gradlew test
Change-Id: I622fda1dafb2ff625224a19f36bbdf1a0877ffc7