Add support for running v4 verification tests from jar files
When the apksig tests are run from gradle, they are run from the
filesystem, so Class#getResource is able to obtain the URI of the
signature file. However, when the tests are run from a jar file,
this fails with an error indicating the URI is not hierarchical.
This commit updates ApkVerifier to obtain the File instance from
the InputStream of the signature file resource in the jar.

This commit also includes minor updates that were reported during
a lint check against previous V4.1 changes.

Bug: 282777109
Test: gradlew test
Change-Id: Ibd9c05375567069c1bfb38cfbc29767bedc2cc67