Add support for multiple error responses during verification
Java 21 changes the behavior of certain error cases during APK
signature verification. One of the existing verification tests
previously failed with an Exception due to an unexpected length,
but as of Java 21, the verification of this signature reports
failure without an exception. This commit updates the signature
verification test to support multiple expected errors to ensure
these tests can run successfully on all supported versions of Java.

Bug: 300311482
Test: gradlew test
Change-Id: I6e6d9620caa6e0bb0e24a27154cde8c47a1ff7f5