Set defaults to align native libraries to 16k
To support page agnostic targets, this commit changes the behavior
of the apksig library and apksigner command to not preserve existing
alignment for files within an APK, but instead native library files
will be aligned on a 16k boundary and all other uncompressed files
will use 4 bytes. APIs and command line options are provided to
maintain the previous behavior by allowing the caller to specify
whether native library alignment should be preserved and if not,
the number of bytes to be used to page align these native library
files.

Bug: 299474707
Test: gradlew test
Change-Id: I5cb455d7ed18f13bd7a5885bd57aa29ec20fee2c