Add apksig-test build target
Run apksig tests via soong (and thus in CI)

This also adds junit.Assume statements in a few
tests:

1. ApkSingerTest#assumeSHA1withDetDSAIsSupported
In the deterministicDsaSigning tests, assumes the
SHA1withDetDSA signature algorithm is supported.
The version of bouncycastle in external/ removes
support for this algorithm, but the upstream
bouncycastle library supports it. Thus, we can
stil run those tests locally.

2. ApkVerifierTest#verify_largeFileSize_doesNotFailWithOOMError
The build server runs these tests with lots of
heap space, so this test was failing with an
ApkFormatException instead. Assume the runtime's
max memory is sufficiently small so the OOM exception
gets thrown and caught.

This change also adds a separate test for
ApkFormatException, and adds the @Test annotation
that was ommitted from one test in
SigningCertificateLineageTest (caught by
errorprone).

Bug: 319296638
Test: atest apksig-test
Change-Id: I406f17e2cfe21993b3a385d54099cef72fd8d647