Use SignerEngine abstraction in ApkSigner
Adds a `KeyConfig` member to the various `SignerConfig` and
corresponding builder classes, so that consumers of the `apksig`
library can pass a configuration to sign *either* via JCA and
a `java.security.PrivateKey`, or via a KMS.

Adds a SignerEngineFactory that fetches a SignerEngine instance
based on the passed config.

Uses this SignerEngineFactory instead of `java.security.Signature`
where relevant.

Deprecates use of a `java.security.PrivateKey` directly in favor
of passing `KeyConfig` objects that can represent either JCA or KMS
signing.

Bug: 319295906
Test: atest apksig-test (existing tests pass)
Change-Id: If3ac5d4c477645edf4846a261513473a529c5998