Use KeyConfig instead of PrivateKey in ApkSignerTool
A previous change deprecates direct references to
`java.security.PrivateKey`s in favor of an internal
`KeyConfig` class that represents signing with either
a locally accessible key and Java Crypto Architecture (JCA),
or a remote key in a Key Management Service (KMS).

Remove usage of the now deprecated ApkSigner APIs in ApkSignerTool.

Bug: 319295906
Test: apksigner sign --key src/test/resources/com/android/apksig/rsa-2048.pk8 --cert src/test/resources/com/android/apksig/rsa-2048.x509.pem --alignment-preserved --in src/test/resources/com/android/apksig/golden-aligned-in.apk --out src/test/resources/com/android/apksig/golden-aligned-out.apk (golden output remains unchanged)
Change-Id: I8dcd2cea3b144b705f2499aee7afcd81a0435823