Use KeyConfig instead of PrivateKey in ApkSigner tests
A previous change deprecates direct references to
`java.security.PrivateKey`s in favor of an internal
`KeyConfig` class that represents signing with either
a locally accessible key and Java Crypto Architecture (JCA),
or a remote key in a Key Management Service (KMS).

Remove usage of the now deprecated ApkSigner APIs in tests.

Bug: 319295906
Test: atest apksig-test (existing tests pass)
Change-Id: Ia01cbbc66d3c63cdcb77dfc12610bf1cf8850017