Add SignerEngine abstraction to support JCA vs. KMS signing
This change adds the unused abstraction `SignerEngine` as well
as an implementation for Java Crytpo Architecture (JCA).

Later commits will introduce the `SignerEngine` by replacing old
calls to `java.security.Signature#sign` with `SignerEngine#sign`.
This provides a simple layer of indirection when we want to sign
with a Key Management Service (e.g. from a cloud provider)
instead of JCA.

Bug: 319295906
Test: N/A (code not yet used)

Change-Id: I70dd254ebeceaea22377da7647b0a957a56c39c2