Register test golden file generation in build.gradle
There is a `main` class in `ApkSignerTest.java` that is used
to generate the "golden" files used in testing. It is only
run locally, to update these files when new ones are added
or something changes.

Add configuration to run this class to `build.gradle` to
better document its usage and expose it at the command line
when running `gradlew`.

Bug: 319296638
Test: ./gradlew createTestGoldens
Change-Id: I55c98b3816fcce087cc5a363b15ad85656e8e69e