Include V3.1 signer(s) in SourceStampVerifier result
The SourceStampVerifier accepts an APK and an SDK range and
obtains all of the V1-V3 signature scheme signers that would be
verified by the platform for the given range. However, support was
never added for the V3.1 signature scheme, so when an APK has a
rotated signer in the V3.1 block and the verifier is called with
an SDK range on which the V3.1 signer would apply, only the original
signer in the V3.0 block is returned. This commit adds support to
the SourceStampVerifier to parse the V3.1 block when present and
return any V3.1 signers that would apply to the specified SDK range.
Note, the source stamp itself does not currently include the V3.1
content digests, so a V3.0 signer will also be returned as the
digest from the V3.0 signer will be used to verify the validity
of the source stamp.

Bug: 329101755
Test: gradlew test
Change-Id: Ic69d62ce301dd119a5d387659110eac9e10cda39