Handle OOM error when allocating space for signature block
When a request is made to append a V2 signature, the existing
V2 signature block needs to be obtained from the APK and a new
block created with the requested signature appended. If the size
of the signature block is larger than the available heap size, an
OutOfMemoryError can be thrown. This commit handles the OOM error
and throws an IOException instead to allow the caller to properly
handle it.

Bug: 287045384
Test: gradlew test
Change-Id: Iffed9a96adee6d9f6d38cbdad269a0e8df302cef