Add BEGIN/END-AOSP comments for KMS tests
apksig gets copied to downstream repos, and we have tests for KMS
integrations that are only supported if the machine running the test has
access to some specific test data in an AWS/GCP account. These comments
allow us to programatically remove bits introduce unnecessary
compile-time dependencies on the relevant KMS libraries that may or may
not be accessible to said downstream repos. This way we don't break
compilation for the sake of tests that probably won't be run downstream
anyway.

Bug: 319296905
Test: TH
Change-Id: I512be49eb51bcd912c5d1ec87260ee45b0f8b35c