Add support for certificate chain in V4 SignerConfig
All previous signature scheme SignerConfigs support specifying a
certificate chain in the SignerConfig as long as the certificate
at the first index corresponds to the current APK signing key.
This commit adds this same support to the V4 signature scheme
by removing the check for multiple certificates in the SignerConfig;
this check is unnecessary since only the first certificate is used,
and it allows the SignerConfig to remain consistent across all
signature scheme signings.

Bug: 356175936
Test: gradlew test
Change-Id: If6eccaefd1d83a5bbdf1f49c86f3a907717b1d99